requirements.txt
streamlit
sqlalchemy
bcrypt
pypdf
python-docx
openpyxl